requirements.txt
flask
python-dotenv
langchain
langchain-groq
groq
google-search-results
gunicorn